fix: address PR review feedback for memo display
- Change MemoInfo.size_bytes from usize to u32 for safer IPC
- Add non-null assertion in getMemoDisplayText
- Add visual feedback for clipboard copy failure
- Remove unused memo_count from transaction query
- Use UNION instead of UNION ALL to deduplicate memos in self-sends
- Fix TransactionInfo fields in job_service for new memo structure

+++ crates/zstash-core/src/domain/transaction.rs
@@ -46,7 +46,7 @@ pub struct MemoInfo {
     /// For Empty memos, this is None.
     pub content: Option<String>,
     /// The size of the memo in bytes (0-512).
-    pub size_bytes: usize,
+    pub size_bytes: u32,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
